chore: rename convert region prices query
