+++ src/Monetization/Application/Query/ConvertRegionPricesQuery.php
@@ -1,10 +1,12 @@
 <?php
 
-namespace Imdhemy\GooglePlay\Monetization\Application;
+declare(strict_types=1);
+
+namespace Imdhemy\GooglePlay\Monetization\Application\Query;
 
 use Imdhemy\GooglePlay\ValueObjects\Money;
 
-final readonly class ConvertRegionPricesPayload
+final readonly class ConvertRegionPricesQuery
 {
     public function __construct(
         public string $packageName,
